Removed commented code from domain class

+++ src/main/java/com/stackroute/muzixapp/service/TrackServiceImpl.java
@@ -1,6 +1,8 @@
 package com.stackroute.muzixapp.service;
 
 import com.stackroute.muzixapp.domain.Track;
+import com.stackroute.muzixapp.exception.TrackAlreadyExistsException;
+import com.stackroute.muzixapp.exception.TrackNotFoundException;
 import com.stackroute.muzixapp.repository.TrackRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -11,7 +13,6 @@ import java.util.Optional;
 @Service
 public class TrackServiceImpl implements TrackService {
     private TrackRepository trackRepository;
-    TrackService trackService = null;
 
     @Autowired
     private TrackServiceImpl(TrackRepository trackRepository) {
@@ -19,34 +20,46 @@ public class TrackServiceImpl implements TrackService {
     }
 
     @Override
-    public Track saveTrack(Track track) {
+    public Track saveTrack(Track track) throws TrackAlreadyExistsException {
+        if (trackRepository.existsById(track.getId())) {
+            throw new TrackAlreadyExistsException("Track already exists!");
+        }
         return trackRepository.save(track);
     }
 
     @Override
-    public List<Track> getAllTracks() {
+    public List<Track> getAllTracks() throws TrackNotFoundException {
+        List<Track> trackList = trackRepository.findAll();
+        if (trackList.isEmpty()) {
+            throw new TrackNotFoundException("Tracks not found");
+        }
         return trackRepository.findAll();
     }
 
     @Override
-    public Track getTrackById(int id) {
+    public Track getTrackById(int id) throws TrackNotFoundException {
+        Optional<Track> trackId = trackRepository.findById(id);
+        if (trackId.isPresent()) {
+            Optional<Track> track_id = trackRepository.findById(id);
+            return track_id.get();
+        } else {
+            throw new TrackNotFoundException("Track not found!");
+        }
+
+    }
+
+    @Override
+    public List<Track> deleteById(int id) throws TrackNotFoundException {
+
         Optional<Track> track_id = trackRepository.findById(id);
-        return track_id.get();
-    }
-
-    @Override
-    public Track deleteTrackById(int id) {
-
+        if (track_id.isEmpty()) {
+            throw new TrackNotFoundException("track not found");
+        }
         trackRepository.deleteById(id);
-        return trackService.getTrackById(id);
+        return trackRepository.findAll();
 
     }
 
-    @Override
-    public List<Track> deleteAllTracks() {
-        trackRepository.deleteAll();
-        return trackService.getAllTracks();
-    }
 
     @Override
     public Track updateById(int id, Track track) {
